+++ backend/src/index.js
@@ -5,8 +5,13 @@ import express from 'express';
 import cors from 'cors';
 import mongoose from 'mongoose';
 import authRoutes from './routes/auth.routes.js';
-import usersRoutes from './routes/user.routes.js';
-import assetsRoutes from './routes/asset.routes.js';
+import userRoutes from './routes/user.routes.js';
+import accessorieRoutes from './routes/accessorie.routes.js';
+import assetRoutes from './routes/asset.routes.js';
+import licenseRoutes from './routes/license.routes.js';
+import locationRoutes from './routes/location.routes.js';
+import companyRoutes from './routes/company.routes.js';
+import departmentRoutes from './routes/department.routes.js';
 import authMiddleware from './middleware/auth.middleware.js';
 
 const app = express();
@@ -18,13 +23,19 @@ app.use(express.json());
 // routes
 app.use('/api/status', (req, res) => res.send('OK'));
 app.use('/api/auth', authRoutes);
-app.use('/api/users', authMiddleware, usersRoutes);
-app.use('/api/assets', authMiddleware, assetsRoutes);
+app.use('/api/users', authMiddleware, userRoutes);
+app.use('/api/accessories', authMiddleware, accessorieRoutes);
+app.use('/api/assets', authMiddleware, assetRoutes);
+app.use('/api/licenses', authMiddleware, licenseRoutes);
+app.use('/api/location', authMiddleware, locationRoutes);
+app.use('/api/company', authMiddleware, companyRoutes);
+app.use('/api/department', authMiddleware, departmentRoutes);
 
 // connections
-mongoose.connect(process.env.MONGO_URI)
+mongoose
+    .connect(process.env.MONGO_URI)
     .then(() => {
         const PORT = process.env.PORT || 3000;
         app.listen(PORT, () => console.log(`App is running on port ${PORT}`));
     })
-    .catch(err => console.error(err));
+    .catch((err) => console.error(err));