+++ backend/src/routes/auth.routes.js
@@ -1,8 +1,7 @@
 import { Router } from 'express';
-import { register, login, logout } from '../controllers/auth.controllers.js';
+import { login, logout } from '../controllers/auth.controllers.js';
 
 const router = Router();
-router.post('/register', register);
 router.post('/login', login);
 router.post('/logout', logout);